Changing the core functionality of the crate
Instead of having to wrap around every panic, this now uses `set_hook` once
to register the callback to print a pretty status message. This also has the
added benefit of pulling in env! calls because the main setup was made
into a macro.

Optionally the two core functions (print_help and dump_log) can now also be used
without the core macro, because they are `pub fn`