Including original panic-cause in report
Okay so this PR is a bit of a dilemma.
The API we need to get the original panic message isn't stable
and requires a feature flag.
We can conditionally enable the feature flag for nightly users
but that doesn't quite deal with the problem of what to do
when someone is using `human-panic` on stable.

As far as I can see there's no other API to get that information?
We could always just gate it all away behind the `nightly` feature
and just ommit a static "not supported" or something when running
on stable.

The tests pass (provided you pass `--features nightly` for now).
Wanted to open this PR to discuss how to proceed :)