Update minimum supported Rust version to 1.32.0
This change bumps the minimum supported Rust version (msrv) to 1.32.0
due to updates in the [backtrace] crate. The update in backtrace was
made in May 2019 ([src][1]) and was later refactored when the CI/build
system was migrated to GitHub Actions ([src][2]). Consequently, either
we'll need to pin the version of backtrace in this crate or ride the
upstream updates and bump the "msrv" accordingly here.

This change assumes that #60 or a similar changeset is also merged which
together should restore the CI to green across the matrix.

[backtrace]: https://crates.io/crates/backtrace
[1]:
https://github.com/rust-lang/backtrace-rs/commit/83744e430271381f0f9a8636fa37fa647820fa45
[2]:
https://github.com/rust-lang/backtrace-rs/blob/816a9559c116f1224c9d3ac968bdc1e6be14fcf1/.github/workflows/main.yml#L196-L205

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>